refactor: rename ConfirmationDialog
