+++ src/collection/components/ConfirmationDialog/ConfirmationDialog.scss
